# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pandac POS API",
        "description": "\n    API for managing products, inventory, sales, transactions, and employees in a POS system.\n\n    **Versioning Strategy**: This API follows semantic versioning. The version is\n    indicated in the base path (`/v1`). Backward-incompatible changes will result\n    in a major version increment.\n    ",
        "termsOfService": "https://pandac.in/terms",
        "contact": {
            "name": "Pandac.in Support",
            "url": "https://pandac.in/",
            "email": "support@pandac.in"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/license/mit"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/sales/": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Read Sales",
                "description": "Retrieve sales with optional filtering.",
                "operationId": "read_sales_v1_sales__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "End Date"
                        }
                    },
                    {
                        "name": "cashier_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Cashier Id"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Customer Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create Sale",
                "description": "Create new sale.",
                "operationId": "create_sale_v1_sales__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaleCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/inventory/stock-levels/": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get Stock Levels",
                "description": "Get stock levels for all products.",
                "operationId": "get_stock_levels_v1_inventory_stock_levels__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "low_stock_only",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Low Stock Only"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/reports/sales": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Sales Report",
                "description": "Get sales report for a date range",
                "operationId": "get_sales_report_v1_reports_sales_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "End Date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Get Sales Report V1 Reports Sales Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/reports/sales/daily": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Daily Sales Report",
                "description": "Get daily sales report for a specific date",
                "operationId": "get_daily_sales_report_v1_reports_sales_daily_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response Get Daily Sales Report V1 Reports Sales Daily Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/reports/sales/period": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Period Sales Report",
                "description": "Get period sales report with flexible parameter names",
                "operationId": "get_period_sales_report_v1_reports_sales_period_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Start"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "End"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "End Date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/reports/inventory/stock-levels": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Inventory Stock Report",
                "description": "Get inventory stock levels report",
                "operationId": "get_inventory_stock_report_v1_reports_inventory_stock_levels_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/v1/reports/customers/loyalty": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Customer Loyalty Report",
                "description": "Get customer loyalty report",
                "operationId": "get_customer_loyalty_report_v1_reports_customers_loyalty_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "top_n",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 50,
                            "title": "Top N"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/reports/financial/profit-loss": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get Profit Loss Report",
                "description": "Get profit and loss report",
                "operationId": "get_profit_loss_report_v1_reports_financial_profit_loss_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Start Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "End Date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SaleCreate": {
                "properties": {
                    "customer_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Customer Id"
                    },
                    "cashier_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Cashier Id"
                    },
                    "sale_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Sale Date"
                    },
                    "subtotal": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Subtotal"
                    },
                    "tax_amount": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Tax Amount"
                    },
                    "discount_amount": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Discount Amount",
                        "default": 0
                    },
                    "total_amount": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Total Amount"
                    },
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SaleStatusEnum"
                            }
                        ],
                        "default": "pending"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 500
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/SaleItemCreate"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Items"
                    }
                },
                "type": "object",
                "required": [
                    "cashier_id",
                    "sale_date",
                    "subtotal",
                    "tax_amount",
                    "total_amount",
                    "items"
                ],
                "title": "SaleCreate"
            },
...
            "SaleItemCreate": {
                "properties": {
                    "product_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Product Id"
                    },
                    "quantity": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Quantity"
                    },
                    "unit_price": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Unit Price"
                    },
                    "subtotal": {
                        "type": "number",
                        "minimum": 0.0,
                        "title": "Subtotal"
                    },
                    "discount_amount": {
                        "anyOf": [
                            {
                                "type": "number",
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Discount Amount",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "product_id",
                    "quantity",
                    "unit_price",
                    "subtotal"
                ],
                "title": "SaleItemCreate"
            },
            "SaleStatusEnum": {
                "type": "string",
                "enum": [
                    "pending",
                    "completed",
                    "cancelled",
                    "refunded"
                ],
                "title": "SaleStatusEnum"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}